mods after unit test

+++ src/main/java/com/ituwei/crud/repo/TodoItemRepo.java
@@ -4,5 +4,5 @@ import com.ituwei.crud.model.TodoItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TodoItemRepo extends CrudRepository<TodoItem,Long> {
+public interface TodoItemRepo extends CrudRepository<TodoItem,Integer> {
 }
